+++ spec/rubocop/cop/style/trailing_comma_in_arguments_spec.rb
@@ -73,6 +73,21 @@ RSpec.describe RuboCop::Cop::Style::TrailingCommaInArguments, :config do
     end
   end
 
+  context 'with a single argument spanning multiple lines' do
+    context 'when EnforcedStyleForMultiline is consistent_comma' do
+      let(:cop_config) { { 'EnforcedStyleForMultiline' => 'consistent_comma' } }
+
+      it 'accepts a single argument with no trailing comma' do
+        expect_no_offenses(<<-RUBY.strip_indent)
+          EmailWorker.perform_async({
+            subject: "hey there",
+            email: "foo@bar.com"
+          })
+        RUBY
+      end
+    end
+  end
+
   context 'with multi-line list of values' do
     context 'when EnforcedStyleForMultiline is no_comma' do
       let(:cop_config) { { 'EnforcedStyleForMultiline' => 'no_comma' } }